loveyouforms-to-node-package
Initial conversion of loveyou-forms 'firebase-project-for-direct-deployment' to
being a node module that can be 'required' from a firebase-project codebase.

Includes updates to 'require' statements of the index.js such as using 'path'
module to include a firbase project's 'service-account.json' file; and
removes use of 'datbaseURL' property of the admin.initializeApp method.